backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.2
ultralytics==8.3.32
Pillow==10.4.0
python-multipart==0.0.20

